requirements.txt
aiofiles==23.2.1
aiosignal==1.2.0
annotated-types==0.7.0
anyio==3.6.2
async-timeout==4.0.2
attrs==22.1.0
beautifulsoup4==4.11.1
boto3==1.34.11
botocore==1.34.162
CacheControl==0.12.11
cachetools==5.2.0
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.5
distro==1.9.0
dnspython==2.2.1
ecdsa==0.18.0
email_validator==2.2.0
fastapi==0.115.4
firebase-admin==6.0.1
frozenlist==1.3.1
gevent==22.10.1
google-api-core==2.10.2
google-api-python-client==2.66.0
google-auth==2.14.1
google-auth-httplib2==0.1.0
google-cloud-core==2.3.2
google-cloud-firestore==2.7.2
google-cloud-storage==2.6.0
google-crc32c==1.5.0
google-resumable-media==2.4.0
googleapis-common-protos==1.57.0
greenlet==1.1.3
grequests==0.6.0
grpcio-status==1.50.0
h11==0.14.0
httpcore==1.0.6
httplib2==0.21.0
httptools==0.5.0
httpx==0.27.2
idna==3.10
itsdangerous==2.1.2
Jinja2==3.1.2
jiter==0.6.1
jmespath==1.0.1
lxml==4.9.1
MarkupSafe==2.1.1
msgpack==1.0.4
multidict==6.0.2
numpy==1.24.1
openai==1.52.0
orjson==3.8.0
pandas==1.5.2
passlib==1.7.4
pillow==10.3.0
playwright==1.27.1
prometheus-fastapi-instrumentator==6.1.0
prometheus_client==0.21.0
proto-plus==1.22.1
protobuf==4.21.9
psycopg2-binary==2.9.10
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pydantic==2.9.2
pydantic-settings==2.6.1
pydantic_core==2.23.4
pydub==0.25.1
pyee==8.1.0
PyJWT==2.6.0
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==0.21.0
python-jose==3.3.0
python-multipart==0.0.5
pytz==2022.7
PyYAML==6.0
requests==2.32.3
rsa==4.9
s3transfer==0.10.3
Shapely==1.8.5.post1
six==1.16.0
sniffio==1.3.0
soupsieve==2.3.2.post1
SpeechRecognition==3.10.4
SQLAlchemy==1.4.42
starlette==0.41.2
tqdm==4.66.5
typing_extensions==4.12.2
ujson==5.5.0
uritemplate==4.1.1
uvicorn==0.18.3
watchfiles==0.18.0
yarl==1.8.1
zope.event==4.5.0
zope.interface==5.5.0
slugify
fastapi-utils
typing-inspect
email-validator==2.1.0
bcrypt==4.0.1